Feature/pn 17312 (#76)
* PN-17312: Add EmdMessageService with implementations and tests

Introduce new EmdMessageService interface to handle message submission functionality.
Includes both enabled and disabled implementations with comprehensive test coverage.

- Add EmdMessageService interface
- Add EmdMessageServiceImpl with full message handling logic
- Add EmdMessageServiceDisabled for when feature is disabled
- Add unit tests for both implementations

* PN-17312: Add EmdRetrievalService with implementations and tests

Introduce new EmdRetrievalService interface to handle retrieval payload operations.
Includes both enabled and disabled implementations with comprehensive test coverage.

- Add EmdRetrievalService interface
- Add EmdRetrievalServiceImpl with retrieval and caching logic
- Add EmdRetrievalServiceDisabled for when feature is disabled
- Add unit tests for both implementations

* PN-17312: Add EmdPaymentService with implementations and tests

Introduce new EmdPaymentService interface to handle payment URL generation.
Includes both enabled and disabled implementations with comprehensive test coverage.

- Add EmdPaymentService interface
- Add EmdPaymentServiceImpl with payment URL creation logic
- Add EmdPaymentServiceDisabled for when feature is disabled
- Add unit tests for both implementations

* PN-17312: Refactor EmdCoreServiceImpl to delegate to specialized services

Simplify EmdCoreServiceImpl by delegating operations to specialized services
(Message, Retrieval, Payment). This improves separation of concerns and makes
the codebase more maintainable.

- Inject EmdMessageService, EmdRetrievalService, and EmdPaymentService
- Delegate submitMessage to EmdMessageService
- Delegate retrieval operations to EmdRetrievalService
- Delegate payment operations to EmdPaymentService
- Remove direct implementation logic from EmdCoreServiceImpl

* PN-17312: Update EmdCoreServiceImpl tests to validate delegation

Refactor tests to focus on validating that EmdCoreServiceImpl correctly
delegates to the specialized services rather than testing implementation details.

- Mock EmdMessageService, EmdRetrievalService, and EmdPaymentService
- Verify delegation to submitMessage
- Verify delegation to retrieval operations
- Verify delegation to payment operations
- Verify error propagation from delegated services

* PN-17312: Update feature flag configuration for granular service control.
Remove deprecated EmdCoreServiceDisabled and its test.
Update conditional tests for new service structure.
