+++ scripts/aws/cfn/application-dev.env
@@ -1,2 +1,4 @@
-#Feature flag che determina se le API esposte dal microservizio devono integrarsi con i servizi esterni.
-PN_EMDINTEGRATION_ISINTEGRATIONENABLED=true
+#Feature flag che determinano se le API esposte dal microservizio devono integrarsi con i servizi esterni.
+PN_EMDINTEGRATION_RETRIEVAL_ENABLED=true
+PN_EMDINTEGRATION_MESSAGE_ENABLED=true
+PN_EMDINTEGRATION_PAYMENT_ENABLED=true
