+++ src/main/java/it/pagopa/pn/emd/integration/service/EmdPaymentService.java
@@ -1,0 +1,8 @@
+package it.pagopa.pn.emd.integration.service;
+
+import it.pagopa.pn.emdintegration.generated.openapi.server.v1.dto.PaymentUrlResponse;
+import reactor.core.publisher.Mono;
+
+public interface EmdPaymentService {
+    Mono<PaymentUrlResponse> getPaymentUrl(String retrievalId, String noticeCode, String paTaxId, Integer amount);
+}
