+++ src/test/java/it/pagopa/pn/emd/integration/service/EmdCoreServiceImplTest.java
@@ -1,19 +1,12 @@
 package it.pagopa.pn.emd.integration.service;
 
-import it.pagopa.pn.emd.integration.cache.AccessTokenExpiringMap;
-import it.pagopa.pn.emd.integration.config.PnEmdIntegrationConfigs;
 import it.pagopa.pn.emd.integration.exceptions.PnEmdIntegrationException;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import it.pagopa.pn.emd.integration.exceptions.PnEmdIntegrationNotFoundException;
-import it.pagopa.pn.emdintegration.generated.openapi.msclient.emdcoreclient.model.RetrievalResponseDTO;
-import it.pagopa.pn.emdintegration.generated.openapi.msclient.milauth.model.AccessToken;
-import it.pagopa.pn.emdintegration.generated.openapi.msclient.emdcoreclient.model.Outcome;
-import it.pagopa.pn.emdintegration.generated.openapi.msclient.emdcoreclient.model.SendMessageRequest;
 import it.pagopa.pn.emdintegration.generated.openapi.msclient.emdcoreclient.model.InlineResponse200;
+import it.pagopa.pn.emdintegration.generated.openapi.msclient.emdcoreclient.model.Outcome;
 import it.pagopa.pn.emdintegration.generated.openapi.server.v1.dto.PaymentUrlResponse;
 import it.pagopa.pn.emdintegration.generated.openapi.server.v1.dto.RetrievalPayload;
 import it.pagopa.pn.emdintegration.generated.openapi.server.v1.dto.SendMessageRequestBody;
-import it.pagopa.pn.emd.integration.middleware.client.EmdClientImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -21,191 +14,87 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
-import java.time.Duration;
-import java.util.Date;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 class EmdCoreServiceImplTest {
 
     @Mock
-    private EmdClientImpl emdClient;
+    private EmdMessageService emdMessageService;
 
     @Mock
-    private AccessTokenExpiringMap accessTokenExpiringMap;
+    private EmdRetrievalService emdRetrievalService;
 
     @Mock
-    private PnEmdIntegrationConfigs pnEmdIntegrationConfigs;
-
-    @Mock
-    private ReactiveRedisService<RetrievalPayload> redisService;
+    private EmdPaymentService emdPaymentService;
 
     @InjectMocks
     private EmdCoreServiceImpl emdCoreServiceImpl;
 
-    PnEmdIntegrationConfigs.CourtesyMessageTemplate digitalMsg = new PnEmdIntegrationConfigs.CourtesyMessageTemplate();
-    PnEmdIntegrationConfigs.CourtesyMessageTemplate analoglMsg = new PnEmdIntegrationConfigs.CourtesyMessageTemplate();
-
-
     @BeforeEach
     void setUp() {
         MockitoAnnotations.openMocks(this);
-
-        digitalMsg.setHeader("Header Digital");
-        digitalMsg.setContent("Contenuto digitale");
-        digitalMsg.setContentFileName("FileContentDigital.md");
-        digitalMsg.setHeaderFileName("FileHeaderDigital.md");
-        analoglMsg.setHeader("Header Analog");
-        analoglMsg.setContent("Contenuto analogico {{schedulingAnalogDate}}");
-        analoglMsg.setContentFileName("FileContentAnalog.md");
-        analoglMsg.setHeaderFileName("FileHeaderAnalog.md");
-
-
-        PnEmdIntegrationConfigs.Templates messagesTemplates = new PnEmdIntegrationConfigs.Templates();
-        messagesTemplates.setAnalogMsg(analoglMsg);
-        messagesTemplates.setDigitalMsg(digitalMsg);
-
-        // 🧩 Mock: quando il service chiede i templates, restituisci il tuo oggetto
-        when(pnEmdIntegrationConfigs.getMsgsTemplates()).thenReturn(messagesTemplates);
     }
 
     @Test
-    void testSubmitMessageAnalog() {
-        // Arrange
+    void testSubmitMessageDelegatesToMessageService() {
         SendMessageRequestBody requestBody = new SendMessageRequestBody();
         requestBody.setRecipientId("recipientId");
-        requestBody.setSenderDescription("senderDescription");
-        requestBody.setAssociatedPayment(true);
-        requestBody.setOriginId("originId");
-        requestBody.setDeliveryMode(SendMessageRequestBody.DeliveryModeEnum.ANALOG);
-        requestBody.setSchedulingAnalogDate( new Date());
-        AccessToken accessToken = new AccessToken();
-        accessToken.setAccessToken("token");
         InlineResponse200 response = new InlineResponse200();
         response.setOutcome(Outcome.OK);
 
-        when(accessTokenExpiringMap.getAccessToken()).thenReturn(Mono.just(accessToken));
-        when(emdClient.submitMessage(any(SendMessageRequest.class),any(String.class),any(String.class))).thenReturn(Mono.just(response));
-        when(pnEmdIntegrationConfigs.getOriginalMessageUrl()).thenReturn("http://example.com");
+        when(emdMessageService.submitMessage(any(SendMessageRequestBody.class))).thenReturn(Mono.just(response));
 
-        // Act
         Mono<InlineResponse200> result = emdCoreServiceImpl.submitMessage(requestBody);
 
-        // Assert
         StepVerifier.create(result)
                 .expectNext(response)
                 .verifyComplete();
+
+        verify(emdMessageService).submitMessage(requestBody);
     }
 
     @Test
-    void testSubmitMessageDigitalWithoutSchedulingDate() {
-        // Arrange
+    void testSubmitMessagePropagatesError() {
         SendMessageRequestBody requestBody = new SendMessageRequestBody();
-        requestBody.setRecipientId("recipientId");
-        requestBody.setSenderDescription("senderDescription");
-        requestBody.setAssociatedPayment(true);
-        requestBody.setOriginId("originId");
-        requestBody.setDeliveryMode(SendMessageRequestBody.DeliveryModeEnum.DIGITAL);
-        AccessToken accessToken = new AccessToken();
-        accessToken.setAccessToken("token");
-        InlineResponse200 response = new InlineResponse200();
-        response.setOutcome(Outcome.OK);
+        PnEmdIntegrationException exception = new PnEmdIntegrationException("Error", 400, "Bad Request");
 
+        when(emdMessageService.submitMessage(any(SendMessageRequestBody.class)))
+                .thenReturn(Mono.error(exception));
 
-        when(accessTokenExpiringMap.getAccessToken()).thenReturn(Mono.just(accessToken));
-        when(emdClient.submitMessage(any(SendMessageRequest.class),any(String.class),any(String.class))).thenReturn(Mono.just(response));
-        when(pnEmdIntegrationConfigs.getOriginalMessageUrl()).thenReturn("http://example.com");
-
-        // Act
         Mono<InlineResponse200> result = emdCoreServiceImpl.submitMessage(requestBody);
 
-        // Assert
         StepVerifier.create(result)
-                    .expectNext(response)
-                    .verifyComplete();
-    }
-    @Test
-    void testSubmitMessageDigitalWithSchedulingDate() {
-        SendMessageRequestBody requestBody = new SendMessageRequestBody();
-        requestBody.setRecipientId("recipientId");
-        requestBody.setSenderDescription("senderDescription");
-        requestBody.setAssociatedPayment(true);
-        requestBody.setOriginId("originId");
-        requestBody.setDeliveryMode(SendMessageRequestBody.DeliveryModeEnum.DIGITAL);
-        requestBody.setSchedulingAnalogDate( new Date());
-        AccessToken accessToken = new AccessToken();
-        accessToken.setAccessToken("token");
-        InlineResponse200 response = new InlineResponse200();
-        response.setOutcome(Outcome.OK);
-
-
-        when(accessTokenExpiringMap.getAccessToken()).thenReturn(Mono.just(accessToken));
-        when(emdClient.submitMessage(any(SendMessageRequest.class),any(String.class),any(String.class))).thenReturn(Mono.just(response));
-        when(pnEmdIntegrationConfigs.getOriginalMessageUrl()).thenReturn("http://example.com");
-
-        // Act
-        Mono<InlineResponse200> result = emdCoreServiceImpl.submitMessage(requestBody);
-
-        // Assert
-        StepVerifier.create(result)
-                    .expectNext(response)
-                    .verifyComplete();
+                .expectError(PnEmdIntegrationException.class)
+                .verify();
     }
 
     @Test
-    void testFailedToSubmitMessageErrorAnalogWithoutSchedulingAnalogDate() {
-        // Arrange
-        SendMessageRequestBody requestBody = new SendMessageRequestBody();
-        requestBody.setRecipientId("recipientId");
-        requestBody.setSenderDescription("senderDescription");
-        requestBody.setAssociatedPayment(true);
-        requestBody.setOriginId("originId");
-        requestBody.setDeliveryMode(SendMessageRequestBody.DeliveryModeEnum.ANALOG);
-        AccessToken accessToken = new AccessToken();
-        accessToken.setAccessToken("token");
-
-        when(accessTokenExpiringMap.getAccessToken()).thenReturn(Mono.just(accessToken));
-        when(emdClient.submitMessage(any(SendMessageRequest.class), any(String.class), any(String.class)))
-                .thenReturn(Mono.error(new RuntimeException("Generic Error")));
-        when(pnEmdIntegrationConfigs.getOriginalMessageUrl()).thenReturn("http://example.com");
-
-        assertThrows(PnEmdIntegrationException.class,
-                     () -> emdCoreServiceImpl.submitMessage(requestBody)
-                    );
-    }
-
-    @Test
-    void getTokenRetrievalPayloadReturnsPayload() {
+    void testGetTokenRetrievalPayloadDelegatesToRetrievalService() {
         String retrievalId = "retrievalId";
-        Boolean isPaymentEnabled = true;
-
         RetrievalPayload expectedPayload = new RetrievalPayload();
         expectedPayload.setRetrievalId(retrievalId);
-        expectedPayload.setIsPaymentEnabled(isPaymentEnabled);
-        RetrievalResponseDTO responseDTO = new RetrievalResponseDTO();
-        responseDTO.setRetrievalId(retrievalId);
-        responseDTO.setIsPaymentEnabled(isPaymentEnabled);
 
-        mockAccessTokenExpiringMap();
-        when(emdClient.getRetrieval(any(String.class), any(String.class))).thenReturn(Mono.just(responseDTO));
-        when(pnEmdIntegrationConfigs.getRetrievalPayloadCacheTtl()).thenReturn(Duration.ofMinutes(5));
-        when(redisService.set(any(String.class), any(RetrievalPayload.class), any(Duration.class))).thenReturn(Mono.empty());
+        when(emdRetrievalService.getTokenRetrievalPayload(retrievalId)).thenReturn(Mono.just(expectedPayload));
 
         Mono<RetrievalPayload> result = emdCoreServiceImpl.getTokenRetrievalPayload(retrievalId);
 
         StepVerifier.create(result)
-                .expectNextMatches(payload -> payload.getRetrievalId().equals(retrievalId) && payload.getIsPaymentEnabled().equals(true) )
+                .expectNext(expectedPayload)
                 .verifyComplete();
+
+        verify(emdRetrievalService).getTokenRetrievalPayload(retrievalId);
     }
 
     @Test
-    void getTokenRetrievalPayloadHandlesNotFound() {
+    void testGetTokenRetrievalPayloadPropagatesError() {
         String retrievalId = "retrievalId";
+        PnEmdIntegrationNotFoundException exception = new PnEmdIntegrationNotFoundException("Error", "Description", "Not Found");
 
-        mockAccessTokenExpiringMap();
-        when(emdClient.getRetrieval(any(String.class), any(String.class)))
-                .thenReturn(Mono.empty());
+        when(emdRetrievalService.getTokenRetrievalPayload(retrievalId))
+                .thenReturn(Mono.error(exception));
 
         Mono<RetrievalPayload> result = emdCoreServiceImpl.getTokenRetrievalPayload(retrievalId);
 
@@ -215,65 +104,30 @@ class EmdCoreServiceImplTest {
     }
 
     @Test
-    void getTokenRetrievalPayloadHandlesError() {
-        String retrievalId = "retrievalId";
-
-        mockAccessTokenExpiringMap();
-        when(emdClient.getRetrieval(any(String.class), any(String.class)))
-                .thenReturn(Mono.error(new RuntimeException("Generic Error")));
-
-        Mono<RetrievalPayload> result = emdCoreServiceImpl.getTokenRetrievalPayload(retrievalId);
-
-        StepVerifier.create(result)
-                .expectErrorMatches(throwable -> throwable instanceof RuntimeException &&
-                        throwable.getMessage().equals("Generic Error"))
-                .verify();
-    }
-
-
-    @Test
-    void getEmdRetrievalPayloadReturnsPayloadFromCache() {
+    void testGetEmdRetrievalPayloadDelegatesToRetrievalService() {
         String retrievalId = "retrievalId";
         RetrievalPayload expectedPayload = new RetrievalPayload();
         expectedPayload.setRetrievalId(retrievalId);
 
-        mockAccessTokenExpiringMap();
-        when(redisService.get(retrievalId)).thenReturn(Mono.just(expectedPayload));
+        when(emdRetrievalService.getEmdRetrievalPayload(retrievalId)).thenReturn(Mono.just(expectedPayload));
 
         Mono<RetrievalPayload> result = emdCoreServiceImpl.getEmdRetrievalPayload(retrievalId);
 
         StepVerifier.create(result)
-                .expectNextMatches(payload -> payload.getRetrievalId().equals(retrievalId))
+                .expectNext(expectedPayload)
                 .verifyComplete();
+
+        verify(emdRetrievalService).getEmdRetrievalPayload(retrievalId);
     }
 
     @Test
-    void getEmdRetrievalPayloadReturnsPayloadFromClient() {
+    void testGetEmdRetrievalPayloadPropagatesError() {
         String retrievalId = "retrievalId";
-        RetrievalResponseDTO responseDTO = new RetrievalResponseDTO();
-        responseDTO.setRetrievalId(retrievalId);
-        RetrievalPayload expectedPayload = new RetrievalPayload();
-        expectedPayload.setRetrievalId(retrievalId);
+        PnEmdIntegrationNotFoundException exception = new PnEmdIntegrationNotFoundException("Error", "Description", "Not Found");
 
-        mockAccessTokenExpiringMap();
-        when(redisService.get(retrievalId)).thenReturn(Mono.empty());
-        when(emdClient.getRetrieval(any(String.class), any(String.class))).thenReturn(Mono.just(responseDTO));
+        when(emdRetrievalService.getEmdRetrievalPayload(retrievalId))
+                .thenReturn(Mono.error(exception));
 
-        Mono<RetrievalPayload> result = emdCoreServiceImpl.getEmdRetrievalPayload(retrievalId);
-
-        StepVerifier.create(result)
-                .expectNextMatches(payload -> payload.getRetrievalId().equals(retrievalId))
-                .verifyComplete();
-    }
-
-    @Test
-    void getEmdRetrievalPayloadHandlesNotFound() {
-        String retrievalId = "retrievalId";
-
-        when(redisService.get(retrievalId)).thenReturn(Mono.empty());
-        when(emdClient.getRetrieval(any(String.class), any(String.class))).thenReturn(Mono.empty());
-
-        mockAccessTokenExpiringMap();
         Mono<RetrievalPayload> result = emdCoreServiceImpl.getEmdRetrievalPayload(retrievalId);
 
         StepVerifier.create(result)
@@ -282,58 +136,40 @@ class EmdCoreServiceImplTest {
     }
 
     @Test
-    void getEmdRetrievalPayloadHandlesError() {
-        String retrievalId = "retrievalId";
-
-        mockAccessTokenExpiringMap();
-        when(redisService.get(retrievalId)).thenReturn(Mono.empty());
-        when(emdClient.getRetrieval(any(String.class), any(String.class)))
-                .thenReturn(Mono.error(new RuntimeException("Generic Error")));
-
-        Mono<RetrievalPayload> result = emdCoreServiceImpl.getEmdRetrievalPayload(retrievalId);
-
-        StepVerifier.create(result)
-                .expectErrorMatches(throwable -> throwable instanceof RuntimeException &&
-                        throwable.getMessage().equals("Generic Error"))
-                .verify();
-    }
-
-    @Test
-    void getPaymentUrlReturnsCorrectUrl() {
+    void testGetPaymentUrlDelegatesToPaymentService() {
         String retrievalId = "retrievalId";
         String noticeCode = "noticeCode";
         String paTaxId = "paTaxId";
         Integer amount = 1000;
-        String emdPaymentEndpoint = "http://example.com/emd_endpoint";
+        PaymentUrlResponse expectedResponse = new PaymentUrlResponse("http://example.com/payment");
 
-        when(pnEmdIntegrationConfigs.getEmdPaymentEndpoint()).thenReturn(emdPaymentEndpoint);
+        when(emdPaymentService.getPaymentUrl(retrievalId, noticeCode, paTaxId, amount))
+                .thenReturn(Mono.just(expectedResponse));
 
         Mono<PaymentUrlResponse> result = emdCoreServiceImpl.getPaymentUrl(retrievalId, noticeCode, paTaxId, amount);
 
         StepVerifier.create(result)
-                .expectNextMatches(response -> response.getPaymentUrl().equals("http://example.com/emd_endpoint?retrievalId=retrievalId&fiscalCode=paTaxId&noticeNumber=noticeCode&amount=1000"))
+                .expectNext(expectedResponse)
                 .verifyComplete();
+
+        verify(emdPaymentService).getPaymentUrl(retrievalId, noticeCode, paTaxId, amount);
     }
 
     @Test
-    void getPaymentUrlReturnsCorrectUrlWithoutAmount() {
+    void testGetPaymentUrlPropagatesError() {
         String retrievalId = "retrievalId";
         String noticeCode = "noticeCode";
         String paTaxId = "paTaxId";
-        String emdPaymentEndpoint = "http://example.com/emd_endpoint";
+        Integer amount = 1000;
+        PnEmdIntegrationException exception = new PnEmdIntegrationException("Error", 400, "Bad Request");
 
-        when(pnEmdIntegrationConfigs.getEmdPaymentEndpoint()).thenReturn(emdPaymentEndpoint);
+        when(emdPaymentService.getPaymentUrl(retrievalId, noticeCode, paTaxId, amount))
+                .thenReturn(Mono.error(exception));
 
-        Mono<PaymentUrlResponse> result = emdCoreServiceImpl.getPaymentUrl(retrievalId, noticeCode, paTaxId, null);
+        Mono<PaymentUrlResponse> result = emdCoreServiceImpl.getPaymentUrl(retrievalId, noticeCode, paTaxId, amount);
 
         StepVerifier.create(result)
-                    .expectNextMatches(response -> response.getPaymentUrl().equals("http://example.com/emd_endpoint?retrievalId=retrievalId&fiscalCode=paTaxId&noticeNumber=noticeCode"))
-                    .verifyComplete();
-    }
-
-    private void mockAccessTokenExpiringMap() {
-        AccessToken accessToken = new AccessToken();
-        accessToken.setAccessToken("token");
-        when(accessTokenExpiringMap.getAccessToken()).thenReturn(Mono.just(accessToken));
+                .expectError(PnEmdIntegrationException.class)
+                .verify();
     }
 }
